Fix `make test` and `make ci` work
The tests have been edited to run directly on the machine hosting the Docker Compose stack, without needing a `runner` container.  This ensures that typing `make test` on the command line works as intended.  This does not affect the current version of the GitHub workflows, which do not use `make`.